Heavily refactored codebase
The model is now split into a seperate inference network and generative model.

The inference network uses an encoder and a normalizing flow (which can be
composed of an arbitrary sequence of maps) to compute the latent state.

The training script now takes a config file instead of command line args.

The config file is copied to the checkpoint directory at the start of training,
and checked for consistency when training is resumed.

Print statements were switched out for logging. There is support to save DEBUG
messages to a text file.

TODO:

- Add functionality for saving/restoring models.
- Refactor `inference.py`.
- Have encoder optionally emit `h` in addition to `mean` and `logv`.
- Implement `CausalCNNMap` in the `normalizing_flows` module.
- Implement `LinearIAFMap` in the `normalizing_flows` module.